requirements.txt
customtkinter==5.1.2
dash==2.7.1
dash_bootstrap_components==1.3.0
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
geopy==2.3.0
numpy==1.24.0
paho_mqtt==1.6.1
pandas==1.5.2
plotly==5.13.1
SQLAlchemy==2.0.7
mysqlclient==2.1.1
Werkzeug==2.2.3
